+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Initial check record downloads script, match hashes again reference files and indirectly verifying DOI resolution
 * Experimental scripts for loading and saving new records from and to a records store (essentially renaming their files) 
 * Documentation on output folder and records store